feat(pi): task-scoped context injection parity with orient --inject
Port PR #184's orient --inject refinements into the Pi extension's
before_agent_start context-injection block. Pi previously dumped raw
JSON.stringify(orient) with a fixed 6000-char truncate and no relevance
gate or opt-out.

- Extract the pure gate + render + config layer out of orient-inject.ts
  into a new dependency-light module orient-inject-render.ts (only
  runtime dep: estimateTokens), so the Pi host can reuse it WITHOUT
  importing the analyzer (handleOrient/tree-sitter) — Pi orients via the
  warm daemon over RPC and must not load the analyzer in-process.
- orient-inject.ts imports + re-exports the moved symbols; the CLI
  orient --inject public surface and behavior are unchanged.
- Pi block C now resolves contextInjection config, honors mode:"off"
  (digest/spec-index baseline grounding unaffected), gates via
  passesRelevanceGate, and renders the token-budgeted, non-coercive
  block via renderInjectionBlock — degrading to the single ignorable
  POINTER_LINE on a weak/absent match instead of dumping verbose JSON.
- CLAUDE.md: add an "MCP tool <-> Pi extension parity" rule.

tsc + eslint clean; orient-inject and pi test suites pass.
